requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
django-webpack-loader==0.7.0
djangorestframework==3.12.1
httpie==2.3.0
idna==2.10
psycopg2==2.8.6
Pygments==2.7.2
PySocks==1.7.1
pytz==2020.1
requests==2.25.0
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
